Refactor DetectDisplayHdrCapable function
Now returns HDREnabled and HDRCapable is a optional bool pointer parameter (pass nullptr if not nedeed).

Note: HDRCapable is no used for now but in a future can be useful to auto-switch Windows HDR ON (if display is HDRcapable and not Enabled).
The code for switch HDR ON in windows exists.